requirements.txt
Pillow>=9.0.0
ImageHash>=4.3
tqdm>=4.60.0
